refactor: extract DEFAULT_SETTINGS to centralized constants file
- Create constants/defaultSettings.ts for default app settings
- Update settingsSlice to import DEFAULT_SETTINGS
- Update SettingsModal to import from constants
- Centralizes default configuration in one location
- Improves maintainability and consistency
